global-entry-appt-notifier/pyproject.toml
[tool.poetry]
name = "global-entry-appt-notifier"
version = "0.1.0"
description = ""
authors = ["Rayshard Thompson <rayshard.thompson@kensho.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
slack-sdk = "^3.15.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
